moved socket.io out from Navbar.js
