+++ client/src/components/Navbar.js
@@ -1,6 +1,6 @@
 import React, { useState, useContext, useEffect } from 'react';
 import PhotoUpload from './PhotoUpload.js';
-import {Toolbar, AppBar, Box, Typography, Button, IconButton, Menu, MenuItem, Avatar, Badge, Popper} from '@material-ui/core';
+import {Toolbar, AppBar, Box, Typography, Button, IconButton, Menu, MenuItem, Avatar} from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { Link as RouterLink } from 'react-router-dom';
 import LocalMallIcon from '@material-ui/icons/LocalMall';
@@ -8,9 +8,7 @@ import AuthContext from '../state_management/AuthContext';
 import { fetchShLists } from '../state_management/actionCreators/shoppingListsActs';
 import ShListsContext from '../state_management/ShListsContext';
 import FindNewFriendsModal from '../components/FindNewFriendsModal.js';
-import socketIOClient from "socket.io-client";
-import Notifications from "./Notifications";
-import {ENDPOINT} from '../utils/baseUrl';
+import SocketContainer from './SocketContainer.js';
 
 
 const useStyles = makeStyles((theme) => ({
@@ -31,24 +29,14 @@ const useStyles = makeStyles((theme) => ({
 	margin: {
 		margin: theme.spacing(1)
 	},
-	popper:{
-		zIndex:1200
-	}
 }));
 
 let needsFetchingLists = true;
-let msgHasBeenRead = false;
-let needsSetSocket = true;
-
 
 const Navbar = (props)=>{
     const classes = useStyles();
     const authContext = useContext(AuthContext);
     const shListsContext = useContext(ShListsContext);
-    const [notification, setNotification] = useState({messages:[]});
-    const [newMsg, setNewMsg] = useState(null);
-    const [socket, setSocket] = useState({socket:null});
-	const [anchorEl, setAnchorEl] = useState(null);
 	const [anchorMenu, setAnchorMenu] = useState(null);
 
 	const handleMenuClick = (event) => {
@@ -65,49 +53,9 @@ const Navbar = (props)=>{
 			needsFetchingLists = false;
 			props.history.push("/main");
         }
-        if(authContext.isAuthenticated && needsSetSocket){
-            const socket = socketIOClient(ENDPOINT);
-            needsSetSocket = false;
 
-            socket.on('show_notification', data => {
-                setNewMsg(data);
-                msgHasBeenRead = false;
-            });
-            socket.emit('join_room', {
-                userId: authContext.id,
-            });
-            setSocket({socket:socket});
-        }
-        if(newMsg){
-			const tmp = [...notification.messages,newMsg].slice(-5);
-            setNotification({ messages:tmp });
-            setNewMsg(null);
-		}
-		// eslint-disable-next-line
-    },[authContext.isAuthenticated, authContext.id, newMsg, shListsContext.dispatchShLists, shListsContext.handleShListsFailure, notification.messages]);
+    },[authContext.isAuthenticated, shListsContext.dispatchShLists, shListsContext.handleShListsFailure]);
 
-    const leaveSocketRoom=()=>{
-        if(socket.socket){
-            socket.socket.emit('leave_room', {
-                userId: authContext.id,
-            });
-        }
-        socket.socket.close();
-        setNotification({ messages:[] });
-        needsSetSocket = true;
-    }
-
-    const handleClickOnNotification = (event) => {
-		if(msgHasBeenRead===true){
-            setNotification({ messages:[] });      
-        }else{
-            msgHasBeenRead = true;
-		}
-        setAnchorEl(anchorEl ? null : event.currentTarget);
-    };
-
-    // const open = Boolean(anchorEl);
-    // const id = open ? 'simple-popper' : undefined;
 
 	return (
 		<AppBar color="primary" elevation={0} className={classes.appBar}>
@@ -136,21 +84,7 @@ const Navbar = (props)=>{
 							Shopping Lists
 						</Button>
 						<FindNewFriendsModal />
-                        <Badge badgeContent={notification.messages.length} color="secondary" overlap="circle">
-						<Button 
-                            aria-describedby='simple-popper'
-                            onClick={handleClickOnNotification}
-                            className={classes.link}>
-                            Notifications
-                            </Button>
-							<Popper 
-								id='simple-popper' 
-								open={Boolean(anchorEl)} 
-								anchorEl={anchorEl} 
-								className={classes.popper}>
-                                <Notifications messages={notification.messages}/>
-                            </Popper>
-                        </Badge>
+                        <SocketContainer/>
 						<IconButton 
 							aria-controls="profile-menu" 
 							aria-haspopup="true" 
@@ -175,7 +109,6 @@ const Navbar = (props)=>{
 							<MenuItem className={classes.link} style={{color: '#DF1B1B'}} onClick={() => {
 									authContext.handleLogout({});
 									shListsContext.handleShListsFailure({response:null});
-                                    leaveSocketRoom();
                                     needsFetchingLists = true;
 									handleMenuClose();
 									}}>
